Code: Fix tm/hm description not being fit
The text now were 2 liners instead of usual 3 or so, it was
not like this when I tested in the vanilla expansion. And the
fact that TM Case is usable, I need to fix this problem.

Credit to ghoulslash for the ReformatItemDescription function,
I took it from his item-desc-header branch and it works well
with extra tweaks added.

Additional changes:
* Expand item desc window width. This technically
  "overlaps" with other window such as the yes/no
  as well with the quantity price window but I've
  made the non-item desc texts using this window
  not exceeding those windows part when they're
  shown. Hacky moment, indeed.
* Update comment in BuyMenuInitGrid
* Remove StringCopy(gStringVar2,
  gMoveNames[ItemIdToBattleMoveId(sShopData->currentItemId)]);
  in line 1384
* Use a const u8 pointer for move name (previously,
  they use gStringVar2 that now clashes with the
  move desc)